Only one company is displayed on map per coordinate
Due the way that geojson data is displayed on map, if there is more than
one geometry point with the same coordinates, only one will be show.

To avoid this problem, we must ensure that:
1 - Every job offer from the same company with the same location receive
the same coordinates.
2 - We offset duplicated coordinates from different companies.

1. We add a database level constrain to ensure that one company have
only one address with the same coordinates.
Raising an 'IntegrityError' (due te above constrain), we know that the
current job offer company already have that address. Therefore we just
need to add the new job to the database address.
If the address doesn't exist, we create a new one and set it to the new
job offer.

At the end of the process, we are sure that there is only one address
per company-coordinates combination. So, every new offer that match this
combination will be add to the current db address. When a company have
more than one job offers 'unfilled', that information will be present on
company map popup as "open positions".
The described logic is made at 'geolocation.geo_utils.add_address_to_job'

2. To be able to display companies with the same coordinates, we
generate a small offset to the coordinates while generating the geojson
data. This create a small performance penalty as we need to check all
coordinates but could be solved by caching the generated geojson. [TODO]
So when we run Map.objects.geojson(), we filter on fly the unfilled
positions, the companies without jobs, and offset duplicated coordinates